requirements.txt
alembic==1.6.5
autopep8==1.5.7
blinker==1.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
flake8==3.9.2
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
Werkzeug==2.0.1
WTForms==2.3.3
